refactor: Update schedule api
This commit contains various changes in the api but for the farmost part it refactors the api endpoints. Here is the detail of all the changes.

1. Restructure api end point (Addressing the feedback from code review). Now using a consistent signal endpoint `/api/graphql/`
2. Move custom graphql nodes to `nodes.py`
3. move mutations to `mutations.py`
4. Remove old code from `booking_schema.py`, `availability_schema.py` and `user_schema.py`.
5. Update urls.py
6. Update Python model test & Graphql API unit tests
7. Create `Makefile` with various actions.
8. Updated the `ReadMe.md` file.
