--- scheduler/api/views.py
@@ -1,1 +1,0 @@
-
