+++ scheduler/meeting_scheduler/apps.py
@@ -1,6 +1,8 @@
+"""Scheduler app config"""
 from django.apps import AppConfig
 
 
 class MeetingSchedulerConfig(AppConfig):
+    """Scheduler app config"""
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'scheduler.meeting_scheduler'
